Created util module
- Move `detect` and `pick` to `util.py`

Resolves #4
